requirements-dev.txt
hiredis==0.2.0
django-nose==1.4
nose==1.3.6
